restructure widgets
