+++ lib/main.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 
 import 'firebase_options.dart';
-import 'package:zakupyapp/screens/manage_lists.dart';
+import 'package:zakupyapp/widgets/manage_lists/manage_lists.dart';
 import 'package:zakupyapp/services/storage_manager.dart';
 import 'package:zakupyapp/utils/app_info.dart';
-import 'package:zakupyapp/screens/home.dart';
-import 'package:zakupyapp/screens/settings.dart';
+import 'package:zakupyapp/widgets/home/home.dart';
+import 'package:zakupyapp/widgets/settings/settings.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
