+++ lib/widgets/home/home.dart
@@ -8,12 +8,12 @@ import 'package:zakupyapp/core/shopping_list_controller.dart';
 import 'package:zakupyapp/core/updater.dart';
 import 'package:zakupyapp/services/storage_manager.dart';
 import 'package:zakupyapp/services/auth_manager.dart';
-import 'package:zakupyapp/utils/snackbars.dart';
+import 'package:zakupyapp/widgets/shared/snackbars.dart';
 import 'package:zakupyapp/widgets/drawer/main_drawer.dart';
 import 'package:zakupyapp/widgets/home/product_card/product_card.dart';
 import 'package:zakupyapp/widgets/shared/full_screen_info.dart';
 import 'package:zakupyapp/widgets/shared/loading.dart';
-import 'package:zakupyapp/widgets/shared/update_dialog.dart';
+import 'package:zakupyapp/widgets/shared/dialogs/update_dialog.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({Key? key}) : super(key: key);
